add patch to make correct merging of booleans possible in deep_merge
The deep_merge gem lacks on a issue, reported here:
https://github.com/danielsdeleo/deep_merge/issues/23
https://github.com/danielsdeleo/deep_merge/pull/28

As yaml_extend uses this gem to merge, here is a workaround to solve
this issue. It will be full forward compatible, if the original deep_merge
solves its issue, due we don't patch deep_merge directly in any way.